put mapquest methods in separate module
